+++ src/main/java/net/mcreator/starcraftvalley/procedures/RevertToDirtProcedure.java
@@ -37,10 +37,9 @@ public class RevertToDirtProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if (world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z)).isSolidSide(world, new BlockPos((int) x, (int) (y + 1), (int) z),
-				Direction.DOWN)) {
+		if (world.getBlockState(new BlockPos(x, y + 1, z)).isSolidSide(world, new BlockPos(x, y + 1, z), Direction.DOWN)) {
 			{
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				BlockState _bs = Blocks.DIRT.getDefaultState();
 				world.setBlockState(_bp, _bs, 3);
 			}
